Insert InfluxDB metrics into tables
Each received line will be parsed and inserted into a table with the
same name as the metric for the line. If the table does not exist, the
line is ignored.

Each table is expected to have the columns `_time`, `_tags`, and
`_fields` in any order.

* The `_time` column need to be of a type that accepts an integer,
  either an 64-bit integer, a timestamp without timezone, or a timestamp
  with timezone.

* The `_tags` and `_fields` columns need to be a JSONB type.

No checking that the fields exist are done.