Support types in parser
Add support to recognize the type of the values and fix the text
representation accordingly. The value is still stored as a pointer into
the buffer, but the trailing "i" for integers is removed in the field
values and plain numerical values are recognized either as decimal
values or integers.