Use prepared statements and separate columns
Instead of building a statement and executing it through the SPI this
commit uses the values directly and prepare a statement with parameters
for execution. It then passes the values to the prepared statement and
execute.

It also add support to insert directly into a column for a tag or a
field, if one exists. The remaining values will be left in `_tags` and
`_fields`. If a key exists in both the tags and the fields, the tag key
will be used.