Fixes to the PGDG workflow
This commit makes some fixes to the previous workflow to ensure that we
can build all supported versions. In addition, we exclude PG15 from the
workflow for now since some header files have moved.

We run the tests inside the official PostgreSQL container to make sure that it
works. This means that we have to start without running the
`docker-entrypoint.sh` script, and will start it later before running the
actual tests.