Add support for spawning background workers on startup
This commit adds support for spawning workers automatically on server startup.
To support this, the extension needs to be loaded using the GUC variable
`shared_preload_libraries` in the configuration file.

The options added in this commit are:

`influx.workers`: The number of workers to spawn. Defaults to 4 workers.

`influx.service`: The service name or port where the workers will listen.
   Defaults to port 8089.

`influx.database`: The name of the database where the workers will connect.

`influx.schema`: The schema name where the workers will write the metrics.

`influx.role`: The role used when writing to the tables. Defaults to the
   superuser.

Closes #9