feat: automatically create tables
When a metric is received that does not have a corresponding table, it will
search for a function named `_create` in the schema where the extension is
installed. If such a function is found, it will be called with the metric, an
array of tags, and an array of field names that was part of the packet. If the
function returns an OID of a relation, that will be used to as target for the
row. If `InvalidOid` is returned, it is assumed that no table is created and
the row will be ignored.