refactor!: move UDP worker to separate file
In preparation to support multiple different protocols we move the UDP-specific
pieces of code to a separate module and keep generic worker code in the worker
module. We also rename some functions to reflect that we will have different
protocols.

BREAKING CHANGE: `worker_launch` is renamed to `launch_udp_worker` and
`send_packet` is renamed to `send_udp_packet`.